Challenges: share is_open()/refusal status and break the import cycle
Move ChallengeSession, CHALLENGE_REFUSAL_STATUS and a shared is_challenge_open()
helper into a new pure leaf module lib/challenge_types.py. models.challenge and
lib.cache.redis now import them at module level instead of via function-local
imports that worked around the lib.tokenclass <-> models circular import;
lib.tokenclass re-exports the names so existing callers are unchanged.

Both challenge backends' is_open() now delegate to the shared helper so they
cannot drift apart, and poll_transaction derives its refusal precedence from
CHALLENGE_REFUSAL_STATUS instead of a hardcoded ("declined", "cancelled") tuple.
