+++ privacyidea/lib/tokenclass.py
@@ -96,6 +96,14 @@ from privacyidea.lib.utils import (is_true, decode_base32check,
                                    to_unicode, create_img, parse_timedelta,
                                    parse_legacy_time, split_pin_pass)
 from .challenge import get_challenges
+from .challenge_types import ChallengeSession
+# Re-exported for backwards compatibility: these live in challenge_types so that
+# models.challenge and lib.cache.redis can import them without a circular import.
+from .challenge_types import (  # noqa: F401
+    CHALLENGE_REFUSAL_STATUS,
+    CHALLENGE_REFUSAL_SESSIONS,
+    is_challenge_open,
+)
 from .config import (get_from_config, get_prepend_pin)
 from .decorators import check_token_locked
 from .error import (TokenAdminError,
@@ -116,20 +124,6 @@ TWOSTEP_DEFAULT_CLIENTSIZE = 8
 TWOSTEP_DEFAULT_DIFFICULTY = 10000
 
 log = logging.getLogger(__name__)
-
-
-class ChallengeSession:
-    ENROLLMENT = "enrollment"
-    DECLINED = "challenge_declined"
-    CANCELLED = "challenge_cancelled"
-
-
-# Maps a refused challenge session to the challenge_status string reported to clients and the
-# audit log. Single source of truth shared by the push answer scan and /validate/polltransaction.
-CHALLENGE_REFUSAL_STATUS = {
-    ChallengeSession.DECLINED: "declined",
-    ChallengeSession.CANCELLED: "cancelled",
-}
 
 
 class Tokenkind:
